+++ src/lib.rs
@@ -783,7 +783,8 @@ pub mod rtp {
     pub use crate::rtp_::{Extension, ExtensionMap, ExtensionSerializer};
 
     pub use crate::packet::{
-        Vp8Descriptor, Vp8DescriptorError, Vp8Patch, Vp8PatchBuilder, Vp8PatchError,
+        RedBlock, RedDecoder, RedEncoder, RedundantBlock, Vp8Descriptor, Vp8DescriptorError,
+        Vp8Patch, Vp8PatchBuilder, Vp8PatchError,
     };
     pub use crate::rtp_::{RtpHeader, SeqNo, Ssrc, VideoOrientation};
     pub use crate::streams::{RtpPacket, RtpWrite, StreamPaused, StreamRx, StreamTx};